migrate to jest test
