+++ package.json
@@ -42,7 +42,6 @@
     "build:watch": "tsc --watch",
     "build": "tsc && shx cp src/*.html dist/src && shx cp src/python/wrapper.py dist/src/python && shx mkdir -p dist/src/golang && shx cp src/golang/wrapper.go dist/src/golang && shx rm -rf dist/drizzle && shx cp -r drizzle dist/drizzle && npm run build:app && shx chmod +x dist/src/main.js",
     "citation:generate": "ts-node scripts/generateCitation.ts",
-    "verify:plugin-sync": "tsx scripts/verify-plugin-sync.ts",
     "db:generate": "npx drizzle-kit generate",
     "db:migrate": "npx tsx src/migrate.ts",
     "db:studio": "npx drizzle-kit studio",
